+++ STACpopulator/implementations/CMIP6_UofT/__init__.py
@@ -1,3 +1,0 @@
-from .add_CMIP6 import CMIP6populator as Populator
-
-__all__ = ["Populator"]