+++ STACpopulator/implementations/CORDEXCMIP6_Ouranos/__init__.py
@@ -1,3 +1,0 @@
-from .add_CORDEX6 import CORDEX_STAC_Populator as Populator
-
-__all__ = ["Populator"]