+++ STACpopulator/implementations/CORDEXCMIP6_Ouranos/add_CORDEX6.py
@@ -10,6 +10,9 @@ LOGGER = logging.getLogger(__name__)
 class CORDEX_STAC_Populator(THREDDSPopulator):
     """Populator that creates STAC objects representing CORDEX data from a THREDDS catalog."""
 
+    name = "CORDEXCMIP6_Ouranos"
+    description = "CMIP6-CORDEX STAC populator from a THREDDS catalog or NCML XML."
+
     data_model = Cordex6DataModel
     item_geometry_model = None  # Unnecessary, but kept for consistency