+++ STACpopulator/implementations/DirectoryLoader/__init__.py
@@ -1,3 +1,0 @@
-from .crawl_directory import DirectoryPopulator as Populator
-
-__all__ = ["Populator"]