+++ STACpopulator/implementations/HRDPS_CRIM/__init__.py
@@ -1,3 +1,0 @@
-from .add_HRDPS import HRDPSpopulator as Populator
-
-__all__ = ["Populator"]