+++ STACpopulator/implementations/RDPS_CRIM/__init__.py
@@ -1,3 +1,0 @@
-from .add_RDPS import RDPSpopulator as Populator
-
-__all__ = ["Populator"]